Consolidate execution hooks
