+++ apps/sim/app/workspace/[workspaceId]/w/[workflowId]/components/panel-new/hooks/index.ts
@@ -1,3 +1,2 @@
 export { usePanelResize } from './use-panel-resize'
-export { useRunWorkflow } from './use-run-workflow'
 export { type UsageData, useUsageLimits } from './use-usage-limits'
